# remote_write: reduce blocking from garbage-collect of series