Fixes #9082 

As suggested at https://github.com/prometheus/prometheus/issues/9082#issuecomment-880691569, 
Add a method `UpdateSeriesSegment()` which is used together with `SeriesReset()` to garbage-collect old series.
This allows us to split the lock around queueManager series data and avoid blocking `Append()` while reading series from the last checkpoint.

Also  `UpdateSeriesSegment()` is massively less expensive than `StoreSeries()` when called with existing series, unless something like #9084 is merged.