@bboreham 2021-07-25
I fixed the comments, and updated the existing `BenchmarkSampleDelivery` to demonstrate the effect.

This benchmark crashed on my machine before I changed it, running out of memory.  It attempted to enqueue 25 million samples in one go, which is not very realistic. I reduced the number of samples and extended the label set on each series to give `StoreSeries()` more work to do.

Results, comparing calling `StoreSeries()` against `UpdateSeriesSegment()`:
```
name              old time/op    new time/op    delta
SampleDelivery-4    88.3ms ± 2%    61.8ms ± 2%  -29.94%  (p=0.008 n=5+5)

name              old alloc/op   new alloc/op   delta
SampleDelivery-4    52.0MB ± 0%    47.2MB ± 0%   -9.29%  (p=0.008 n=5+5)

name              old allocs/op  new allocs/op  delta
SampleDelivery-4      425k ± 0%      414k ± 0%   -2.56%  (p=0.008 n=5+5)
```