@bboreham 2021-07-27
> the current benchmark should still call the StoreSeries function 

It does, before the timing loop, just as it did before this PR.
Are you suggesting that `StoreSeries` should be called inside the timing loop?

Can you describe the real-life scenario you want the benchmark to imitate?