@bboreham 2021-07-27
>  I am curious how much of the current benchmark is spent processing samples vs running the gc?

Removing the two lines calling `UpdateSeriesSegment` and `SeriesReset` makes negligible difference.

```
$ benchstat nogc.txt withgc.txt 
name              old time/op    new time/op    delta
SampleDelivery-4    61.1ms ± 4%    60.7ms ± 3%   ~     (p=0.841 n=5+5)

name              old alloc/op   new alloc/op   delta
SampleDelivery-4    47.1MB ± 0%    47.2MB ± 0%   ~     (p=0.286 n=4+5)

name              old allocs/op  new allocs/op  delta
SampleDelivery-4      414k ± 0%      414k ± 0%   ~     (p=0.333 n=4+5)
```