@csmarchbanks 2021-07-27
Thanks, I am happy with this as is. I will let Callum respond and merge if he is happy as well.