@cstyan 2021-07-27
> It does, before the timing loop, just as it did before this PR.

I must have misread something then, LGTM

> Are you suggesting that StoreSeries should be called inside the timing loop?

I think it would be worth exploring a benchmark between the two, but that's not necessary for this PR.